refactor: change folder and file structure
